rearrange stuff and fix lint problems
