+++ functions/src/admin-functions/newUser.ts
@@ -1,6 +1,6 @@
 import { onValueCreated } from "firebase-functions/database";
-import { GmailConfig, sendEmail } from "../mail/gmail";
-const { defineSecret } = require("firebase-functions/params");
+import { GmailConfig, sendEmail } from "../gmail";
+import { defineSecret } from "firebase-functions/params";
 
 // Define the secret parameter
 const gEmail = defineSecret("GMAIL_EMAIL");
@@ -14,27 +14,26 @@ type Userdata = {
 }
 
 const sendNewUserNotificationEmail = onValueCreated({ ref: "/users/{userId}", secrets: [gEmail, gPass] }, (event) => {
-
   const original = event.data.val();
-  const userdata = { ...original }
-  const userId = event.params.userId
+  const userdata = { ...original };
+  const userId = event.params.userId;
 
-  onNewUser(userId, userdata)
+  onNewUser(userId, userdata);
 });
 
 const onNewUser = (userId: string, userdata: Userdata) => {
-  console.log("notifyNewUser: ", userId, userdata)
+  console.log("notifyNewUser: ", userId, userdata);
 
   const gmailConfig: GmailConfig = {
     email: gEmail.value(),
-    password: gPass.value()
+    password: gPass.value(),
   };
 
   const to = "fest@nightofpassion.se";
 
-  const mailContent = onNewUserEmailTemplate(userdata)
-  return sendEmail(gmailConfig, to, "New NoP signup", mailContent)
-}
+  const mailContent = onNewUserEmailTemplate(userdata);
+  return sendEmail(gmailConfig, to, "New NoP signup", mailContent);
+};
 
 type NewUserSignup = {
   username: string,
@@ -46,7 +45,7 @@ const onNewUserEmailTemplate = (data: NewUserSignup): string => {
 user: ${data.username || "<unknown user>"}    
 message: ${data.message || "<unknown msg>"}     
 `;
-  return text
-}
+  return text;
+};
 
 export { sendNewUserNotificationEmail };
